CCTM 3.0.0.0; Add: Can log return value of stored procedures; Add: Can log enumeration arguments/returns for stored procedures; Add: Migrated the following procedures to the new database system: UpdTransactionauthorization, UpdTransactionrecordStatus, UpdTransactionauthorizationStatus, SelNewTransactionrecords, UpdTransactionrecordCctm;
AutoDatabase:
Mod: IDatabaseMethodTracker can now intercept query results;

git-svn-id: https://nas.ipsgroupinc.com:88/svn/ParkingMeter/trunk/TransactionManagement@717 1cffc995-40da-c24f-b319-404966fbbd81

+++ TransactionManagementCommonUnitTests/RtccUnitTests.cs
@@ -110,13 +110,13 @@ namespace UnitTests
 
         private byte[] ipsEncryptStripe(string stripe, TransactionManagementCommon.TransactionInfo info, int keyVersion)
         {
-            return CryptographicPlatforms.IPSTrackCipher.Encrypt(stripe, info.StartDateTime, info.TransactionIndex,
+            return CryptographicPlatforms.IPSTrackCipher.Encrypt(stripe, info.StartDateTime, (int)info.TransactionIndex,
                 Int32.Parse(info.MeterSerialNumber), keyVersion);
         }
 
         private UnencryptedStripe ipsDecryptStripe(EncryptedStripe stripe, TransactionManagementCommon.TransactionInfo info, int keyVersion)
         {
-            return CryptographicPlatforms.IPSTrackCipher.Decrypt(stripe, info.RefDateTime??info.StartDateTime, info.TransactionIndex,
+            return CryptographicPlatforms.IPSTrackCipher.Decrypt(stripe, info.RefDateTime??info.StartDateTime, (int)info.TransactionIndex,
                 Int32.Parse(info.MeterSerialNumber), keyVersion);
         }
